수정: String Extension 제거
