+++ Projects/Core/Models/Network/MyPage/ResponseDTO/TodayShortsResponseDTO.swift
@@ -28,3 +28,16 @@ public extension TodayShortsResponseDTO {
     )
   }
 }
+
+fileprivate extension String {
+  func toDate() -> Date {
+    let dateFormatter = DateFormatter()
+    dateFormatter.dateFormat = "yyyy-MM-ddTHH:mm:ss"
+    dateFormatter.timeZone = TimeZone(identifier: "UTC")
+    
+    if let date = dateFormatter.date(from: self) {
+      return date
+    }
+    return Date()
+  }
+}
